app/requirements.txt
click==8.1.6
flake8==6.0.0
Flask==2.2.5
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
gunicorn==21.2.0
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
packaging==23.1
pycodestyle==2.10.0
pyflakes==3.0.1
PyMySQL==1.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
Werkzeug==2.3.6
WTForms==3.0.1
wtforms-html5==0.6.1
